refactor: share the contact between rides and lost items
