+++ FateConnect/Web/src/components/ContactButton/constants/index.ts
@@ -1,0 +1,7 @@
+export const CONTACT_LABEL = 'Contato';
+
+export const CONTACT_DIALOG = {
+  title: 'Informações de Contato',
+  emailCopied: 'E-mail copiado!',
+  emailCopyFailed: 'Não foi possível copiar o e-mail.',
+};
